Move KubernetesResources to appgw
