+++ pkg/appgw/backendhttpsettings.go
@@ -9,7 +9,6 @@ import (
 	"errors"
 	"fmt"
 	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/annotations"
-	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/k8scontext"
 	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-12-01/network"
 	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/golang/glog"
@@ -196,7 +195,7 @@ func (c *appGwConfigBuilder) getBackendsAndSettingsMap(ingressList []*v1beta1.In
 	return &httpSettings, backendHTTPSettingsMap, finalServiceBackendPairMap, nil
 }
 
-func (c *appGwConfigBuilder) BackendHTTPSettingsCollection(kr *k8scontext.KubernetesResources) error {
+func (c *appGwConfigBuilder) BackendHTTPSettingsCollection(kr *KubernetesResources) error {
 	httpSettings, _, _, err := c.getBackendsAndSettingsMap(kr.IngressList, kr.ServiceList)
 	c.appGwConfig.BackendHTTPSettingsCollection = httpSettings
 	return err
