+++ pkg/k8scontext/types.go
@@ -2,13 +2,8 @@ package k8scontext
 
 import (
 	"github.com/eapache/channels"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/api/extensions/v1beta1"
 	"k8s.io/client-go/tools/cache"
 
-	mtv1 "github.com/Azure/application-gateway-kubernetes-ingress/pkg/apis/azureingressmanagedtarget/v1"
-	ptv1 "github.com/Azure/application-gateway-kubernetes-ingress/pkg/apis/azureingressprohibitedtarget/v1"
-	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/environment"
 	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/utils"
 )
 
@@ -46,14 +41,4 @@ type Context struct {
 	stopChannel       chan struct{}
 
 	UpdateChannel *channels.RingChannel
-}
-
-// KubernetesResources holds the structs we have fetches from Kubernetes + environment, based on which
-// we will construct App Gateway config.
-type KubernetesResources struct {
-	IngressList       []*v1beta1.Ingress
-	ServiceList       []*v1.Service
-	ManagedTargets    []*mtv1.AzureIngressManagedTarget
-	ProhibitedTargets []*ptv1.AzureIngressProhibitedTarget
-	EnvVariables      environment.EnvVariables
 }
